refactor: code cleaning and error fixing
