+++ OrganiTask/Controllers/DashboardController.cs
@@ -346,7 +346,7 @@ namespace OrganiTask.Controllers
                 //        // Luego, eliminamos las relaciones entre las etiquetas y las tareas
                 //        foreach (int tagId in tagIds)
                 //        {
-                //            DeleteTag(tagId); // Eliminar cada etiqueta asociada a la categoría
+                //            new TagController.DeleteTag(tagId); // Eliminar cada etiqueta asociada a la categoría
                 //        }
 
                 //        // Ahora eliminamos la categoría en sí
@@ -381,183 +381,6 @@ namespace OrganiTask.Controllers
 
                 // Si no se encuentra la categoría, retornamos un modelo vacío
                 return cat ?? new CategoryViewModel { Id = categoryId, Title = string.Empty };
-            }
-        }
-
-        /// <summary>
-        /// Obtiene las etiquetas asociadas a una categoría específica.
-        /// </summary>
-        /// <param name="categoryId">Identificador de la categoría.</param>
-        /// <returns>Lista de etiquetas asociadas a la categoría.</returns>
-        public OrganiList<TagViewModel> GetTagsForCategory(int categoryId)
-        {
-            // Usamos un bloque using para asegurarnos de que el contexto se libere al finalizar
-            using (OrganiTaskDB context = new OrganiTaskDB())
-            {
-                return context.Tags
-                    .Where(t => t.CategoryId == categoryId)
-                    .Select(t => new TagViewModel
-                    {
-                        Id = t.Id,
-                        Name = t.Name,
-                        Color = t.Color,
-                        CategoryId = t.CategoryId
-                    })
-                    .ToOrganiList();
-            }
-        }
-
-        /// <summary>
-        /// Crea una nueva etiqueta y la asocia a una categoría.
-        /// </summary>
-        /// <param name="tagVm">Modelo de vista de la etiqueta a crear.</param>
-        public void CreateTag(TagViewModel tagVm)
-        {
-            using (OrganiTaskDB context = new OrganiTaskDB())
-            {
-                // Creamos el objeto Tag y lo añadimos al contexto
-                Tag tag = new Tag
-                {
-                    Name = tagVm.Name,
-                    Color = tagVm.Color,
-                    CategoryId = tagVm.CategoryId
-                };
-                context.Tags.Add(tag);
-                context.SaveChanges(); // Guardamos los cambios en la base de datos
-                tagVm.Id = tag.Id; // Asignamos el ID generado a la vista modelo
-            }
-        }
-
-        /// <summary>
-        /// Actualiza los datos de una etiqueta existente.
-        /// </summary>
-        /// <param name="updateTag">Modelo de vista de la etiqueta a actualizar.</param>
-        public void UpdateTag(TagViewModel updateTag)
-        {
-            // Usamos un bloque using para asegurarnos de que el contexto se libere al finalizar
-            using (OrganiTaskDB context = new OrganiTaskDB())
-            {
-                // Buscamos la etiqueta por su ID
-                Tag tag = context.Tags.FirstOrDefault(t => t.Id == updateTag.Id);
-                if (tag == null) return; // Si no se encuentra la etiqueta, no hacemos nada
-
-                // Actualizamos los datos de la etiqueta
-                tag.Name = updateTag.Name;
-                tag.Color = updateTag.Color;
-                context.SaveChanges(); // Guardar los cambios
-            }
-        }
-
-        /// <summary>
-        /// Elimina una etiqueta y sus relaciones con las tareas.
-        /// </summary>
-        /// <param name="tagId">Identificador de la etiqueta a eliminar.</param>
-        public void DeleteTag(int tagId)
-        {
-            // Usamos un bloque using para asegurarnos de que el contexto se libere al finalizar
-            using (OrganiTaskDB context = new OrganiTaskDB())
-            {
-                // Buscamos las relaciones entre la etiqueta y las tareas
-                OrganiList<TaskTag> links = context.TaskTags.Where(tt => tt.TagId == tagId).ToOrganiList();
-                if (links.Any())
-                    context.TaskTags.RemoveRange(links); // Eliminar los enlaces de la etiqueta a las tareas
-
-                // Una vez eliminada las relaciones, eliminamos la etiqueta
-                Tag tag = context.Tags.FirstOrDefault(t => t.Id == tagId);
-                if (tag != null)
-                    context.Tags.Remove(tag); // Eliminar la etiqueta
-
-                context.SaveChanges(); // Guardar los cambios
-            }
-        }
-
-        /// <summary>
-        /// Crea una nueva categoría asociada a un tablero específico.
-        /// </summary>
-        /// <param name="category">Modelo de vista de la categoría a crear.</param>
-        public void CreateCategory(CategoryViewModel category)
-        {
-            using (OrganiTaskDB context = new OrganiTaskDB())
-            {
-                Category cat = new Category
-                {
-                    Title = category.Title,
-                    DashboardId = category.DashboardId
-                };
-                context.Categories.Add(cat);
-                context.SaveChanges();
-                category.Id = cat.Id;
-            }
-        }
-
-        /// <summary>
-        /// Actualiza los datos de una categoría existente.
-        /// </summary>
-        /// <param name="category">Modelo de vista de la categoría a actualizar.</param>
-        public void UpdateCategory(CategoryViewModel category)
-        {
-            using (OrganiTaskDB context = new OrganiTaskDB())
-            {
-                Category cat = context.Categories.FirstOrDefault(c => c.Id == category.Id);
-
-                if (cat == null) return;
-
-                cat.Title = category.Title;
-                context.SaveChanges();
-            }
-        }
-
-        /// <summary>
-        /// Elimina una categoría y todas las etiquetas asociadas a ella.
-        /// </summary>
-        /// <param name="categoryId">Identificador de la categoría a eliminar.</param>
-        public void DeleteCategory(int categoryId)
-        {
-            using (OrganiTaskDB context = new OrganiTaskDB())
-            {
-                // Primero obtenemos todas las etiquetas asociadas a la categoría
-                OrganiList<int> tagIds = context.Tags
-                    .Where(t => t.CategoryId == categoryId)
-                    .Select(t => t.Id)
-                    .ToOrganiList();
-
-                // Luego, eliminamos las relaciones entre las etiquetas y las tareas
-                foreach (int tagId in tagIds)
-                {
-                    DeleteTag(tagId); // Eliminar cada etiqueta asociada a la categoría
-                }
-
-                // Ahora eliminamos la categoría en sí
-                Category category = context.Categories.FirstOrDefault(c => c.Id == categoryId);
-                if (category != null)
-                {
-                    context.Categories.Remove(category); // Eliminar la categoría
-                    context.SaveChanges(); // Guardar los cambios
-                }
-            }
-        }
-
-        /// <summary>
-        /// Verifica si una categoría específica existe en el tablero
-        /// </summary>
-        /// <param name="dashboardId">ID del tablero</param>
-        /// <param name="categoryId">ID de la categoría a verificar</param>
-        /// <returns>True si la categoría existe, False en caso contrario</returns>
-        public bool CategoryExists(int dashboardId, int categoryId)
-        {
-            try
-            {
-                // Obtener todas las categorías del tablero
-                OrganiList<CategoryViewModel> categories = GetDashboardCategories(dashboardId);
-
-                // Verificar si la categoría con el ID específico existe
-                return categories != null && categories.Any(c => c.Id == categoryId);
-            }
-            catch (Exception ex)
-            {
-               MessageBox.Show($"Error al verificar si existe la categoría: {ex.Message}",
-                   "Hubo un error", MessageBoxButtons.OK, MessageBoxIcon.Error);
-                return false;
             }
         }
     }
